+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2025-12-19:
+  - updated packages.
+  - added language parameter to build word index command.
+  - added ignore POS mismatches to word index check command.
 - 2025-12-18:
   - updated packages.
   - added whitelist to word checker.